AbstractJarFileLocator -> AbstractJarFileModLocator
This creates a hard dependency on certain FML versions. For now this is the quickest fix to get it working again. Need to roll back to implement IModLocator directly if we ever consider max-range compatibility.
Bump version to 0.2.4. When I get time to roll back to IModLocator directly, this will bump to 0.3.0

Fix GH-3